move two modules under one file
